+++ src/Crypto/Signature/Signature.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Mdanter\Ecc\Crypto\Routines\Signature;
+namespace Mdanter\Ecc\Crypto\Signature;
 
 /**
  * *********************************************************************
@@ -27,21 +27,49 @@ namespace Mdanter\Ecc\Crypto\Routines\Signature;
  */
 
 /**
- * This is the contract for describing a signature used in ECDSA.
+ * Plain Old PHP Object that stores the signature r,s for ECDSA
  */
-interface SignatureInterface
+class Signature implements SignatureInterface
 {
     /**
-     * Returns the r parameter of the signature.
      *
-     * @return int|string
+     * @var int|string
      */
-    public function getR();
+    protected $r;
 
     /**
-     * Returns the s parameter of the signature.
      *
-     * @return int|string
+     * @var int|string
      */
-    public function getS();
+    protected $s;
+
+    /**
+     * Initialize a new instance with values
+     *
+     * @param int|string $r
+     * @param int|string $s
+     */
+    public function __construct($r, $s)
+    {
+        $this->r = $r;
+        $this->s = $s;
+    }
+
+    /**
+     * {@inheritDoc}
+     * @see \Mdanter\Ecc\SignatureInterface::getR()
+     */
+    public function getR()
+    {
+        return $this->r;
+    }
+
+    /**
+     * {@inheritDoc}
+     * @see \Mdanter\Ecc\SignatureInterface::getS()
+     */
+    public function getS()
+    {
+        return $this->s;
+    }
 }
